- Start on LangProvider - Fixed Lasers not doing intended damage - Mostly finished Armour - Fixed that stupid skin bad (although in a didgy way, too bad!) - Fixed broken trait translations - Removed Lucky trait (literally never worked and don't know why it was still here) - Fixed biomes making new colors for their grass constantly

+++ src/main/java/me/swirtzly/regeneration/common/item/GunItem.java
@@ -24,9 +24,9 @@ import net.minecraft.world.World;
 public class GunItem extends SolidItem {
 
     private final int cooldown;
-    private final int damage;
+    private final float damage;
 
-    public GunItem(int shotsPerRound, int cooldown, int damage) {
+    public GunItem(int shotsPerRound, int cooldown, float damage) {
         super(new Properties().group(ItemGroups.REGEN_TAB).maxDamage(shotsPerRound).setNoRepair());
         this.cooldown = cooldown;
         this.damage = damage;
@@ -83,7 +83,7 @@ public class GunItem extends SolidItem {
         }
     }
 
-    public int getDamage() {
+    public float getDamage() {
         return damage;
     }
